databases/postgres/requirements.txt
pandas
polars
psycopg2-binary